% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strandr.R
\name{strandr}
\alias{strandr}
\title{strandr}
\usage{
strandr(table_input = "", lat, lon)
}
\arguments{
\item{table_input}{a table to which the returned data will be added}

\item{lat}{a numerical value, or vector of numerical values, that specifies the latitude(s) in the input coordinates}

\item{lon}{a numerical value, or vector of numerical values, that specifies the longitude(s) in the input coordinates}
}
\value{
dataframe with Strand-codes and locality data
}
\description{
Returns Strand-codes and locality data from one, or a set of input coordinates (projection longlat, datum WGS84). The strand-codes are obtained by identifying the intesects beween the input coordinates and a polygon dataset of the Strand regions. Locality names are obtained from the Norwegian locality-name-base (Sentralt Stedsnavnregister). For each input coordinate, the function returns the locality name  that is closest in distance. The maximum distance from a locality name is 2000 m.
}
\examples{
example_input <- data.frame("COL_ID" = c("JPL_0051", "JPL_0052", "JPL_0053"), 
                            "Latitude" = c(69.57696, 70.44070, 62.259262),
                            "Longitude" = c(24.840064, 23.186622, 12.734821))
example_output <- strandr(example_input, 
                          lat = example_input$Latitude, 
                          lon = example_input$Longitude)
}
