% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Strandkoder.R
\name{Strandkoder}
\alias{Strandkoder}
\title{Returns strand-codes}
\usage{
Strandkoder(z, long, lat)
}
\arguments{
\item{z}{(dataframe) A table containing coordinates. Georeference system must be longlat WGS84/Euref89. Longitudes and latitudes must be organized in separate rows. Whether there are additional rows of data in the dataset makes no difference.}

\item{long}{The specific row in the input table containing longitudes}

\item{lat}{The specific row in the input table containing latitudes}
}
\value{
Returns the input dataframe with strand-codes added to it
}
\description{
Adds strand-codes to an input table of coordinates. The strand-codes are obtained by identifying the intesects beween the input coordinates and a polygon dataset of the Norwegian municipality borders (Anno 2012)
}
\examples{
ex_in <- data.frame("COL_ID" = c("JPL0051", "JPL0052"), "Longitude" =
         c(24.840064, 23.186622), "Latitude" = c(69.57696, 70.44070))
ex_out <- Strandkoder(ex_in, long = ex_in$Longitude, lat = ex_in$Latitude)
}
